refactor: moved conversion related stuff to filter since it might be dependent on the selected filter
